requirements.txt
aioredis==1.2.0
asgi-redis==1.4.3
asgiref==1.1.2
async-timeout==2.0.1
attrs==18.2.0
autobahn==18.11.1
Automat==0.7.0
certifi==2018.10.15
channels==1.1.8
channels-api==0.4.1
chardet==3.0.4
constantly==15.1.0
daphne==1.4.2
Django==2.1.3
djangorestframework==3.9.0
gunicorn==19.7.1
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
msgpack==0.5.6
msgpack-python==0.5.6
mysqlclient==1.3.14
PyHamcrest==1.9.0
PyMySQL==0.9.3
pytz==2018.7
redis==2.10.6
requests==2.20.1
six==1.11.0
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.1
zope.interface==4.6.0
django-crispy-forms==1.7.2
django-recaptcha==2.0.2
uwsgi==2.0.18


